Updated Tensorflow syntax and fixed accuracy value
Much of the syntax used in this file for Tensorflow has been deprecated. The syntax updated here will allow for use with any other possible legacy code. Additionally Keras syntax needed to be updated from `K.set_session(sess)` to `tf.compat.v1.keras.backend.set_session(sess)`. Finally the use of 'acc' in many instances of the code in this file is unusable, as is related to issue #4 on this repo. 

Without these changes someone would need to look up the broken syntax on their own and correct it - perhaps that's the point, however it wasn't clearly stated in this exercise that was part of the intent.


https://github.com/tensorflow/tensorflow/issues/33504
https://github.com/tensorflow/tensorflow/issues/18538
https://github.com/keras-team/keras/issues/12379
https://github.com/tensorflow/tensorflow/issues/30070
https://www.tensorflow.org/api_docs/python/tf/compat/v1/keras/backend/set_session